requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
django-templated-mail==1.1.1
djangorestframework==3.10.2
djangorestframework-jsonapi==2.8.0
djangorestframework-jwt==1.11.0
djoser==2.0.1
inflection==0.3.1
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
